feat(scaffold): lift FpsOverlay to shell-owned (#261)
Move `FpsOverlay` allocation, scene-graph parenting, animation-loop
tick, and dispose into `src/shell/shell.ts`. Delete the four
near-identical per-scene wirings introduced by the #264 stopgap so
future scenes (#233 capability audit, post-1.0 portfolio) get the
`?fps=1` overlay for free.

`RendererInfoProbe` (#102) stays quadrics-local under the same
`?fps=1` gate — only that scene has the raymarcher fragment-cost
question the probe was added to answer. The two `isFpsOverlayEnabled`
copies are intentional and documented.

Disposer pushed AFTER the renderer disposer so the LIFO drain runs
`fpsOverlay.dispose()` (releases troika Text GL resources) BEFORE
`renderer.dispose()` frees the GL context — mirrors the
environment-disposer pattern from #224.
